debt(GraphQL): prevent importing V1 types from V2
