+++ server/graphql/v1/CollectiveInterface.js
@@ -16,8 +16,7 @@ import sequelize from 'sequelize';
 import SqlString from 'sequelize/lib/sql-string';
 
 import { types } from '../../constants/collectives';
-import FEATURE, { FeaturesList } from '../../constants/feature';
-import FEATURE_STATUS from '../../constants/feature-status';
+import FEATURE from '../../constants/feature';
 import { PAYMENT_METHOD_SERVICE, PAYMENT_METHOD_TYPE } from '../../constants/paymentMethods';
 import roles from '../../constants/roles';
 import { isCollectiveDeletable } from '../../lib/collectivelib';
@@ -26,9 +25,8 @@ import queries from '../../lib/queries';
 import { canSeeLegalName } from '../../lib/user-permissions';
 import models, { Op } from '../../models';
 import { hostResolver } from '../common/collective';
+import { CollectiveFeatures } from '../common/CollectiveFeatures';
 import { getContextPermission, PERMISSION_TYPE } from '../common/context-permissions';
-import { getFeatureStatusResolver } from '../common/features';
-import { getIdEncodeResolver, IDENTIFIER_TYPES } from '../v2/identifiers';
 import { Policies } from '../v2/object/Policies';
 import { SocialLink } from '../v2/object/SocialLink';
 
@@ -881,19 +879,6 @@ export const CollectiveInterfaceType = new GraphQLInterfaceType({
     };
   },
 });
-
-const FeaturesFields = () => {
-  return FeaturesList.reduce(
-    (obj, feature) =>
-      Object.assign(obj, {
-        [feature]: {
-          type: CollectiveFeatureStatus,
-          resolve: getFeatureStatusResolver(feature),
-        },
-      }),
-    {},
-  );
-};
 
 const CollectiveFields = () => {
   return {
@@ -1964,39 +1949,6 @@ const CollectiveFields = () => {
     },
   };
 };
-
-export const CollectiveFeatureStatus = new GraphQLEnumType({
-  name: 'CollectiveFeatureStatus',
-  values: {
-    [FEATURE_STATUS.ACTIVE]: {
-      description: 'The feature is enabled and is actively used',
-    },
-    [FEATURE_STATUS.AVAILABLE]: {
-      description: 'The feature is enabled, but there is no data for it',
-    },
-    [FEATURE_STATUS.DISABLED]: {
-      description: 'The feature is disabled, but can be enabled by an admin',
-    },
-    [FEATURE_STATUS.UNSUPPORTED]: {
-      description: 'The feature is disabled and cannot be activated for this account',
-    },
-  },
-});
-
-export const CollectiveFeatures = new GraphQLObjectType({
-  name: 'CollectiveFeatures',
-  description: 'Describes the features enabled and available for this account',
-  fields: () => {
-    return {
-      id: {
-        type: new GraphQLNonNull(GraphQLString),
-        description: 'The id of the account',
-        resolve: getIdEncodeResolver(IDENTIFIER_TYPES.ACCOUNT),
-      },
-      ...FeaturesFields(),
-    };
-  },
-});
 
 export const CollectiveType = new GraphQLObjectType({
   name: 'Collective',
