+++ server/graphql/v2/input/UpdateCreateInput.js
@@ -1,6 +1,5 @@
 import { GraphQLBoolean, GraphQLInputObjectType, GraphQLNonNull, GraphQLString } from 'graphql';
-
-import { IsoDateString } from '../../v1/types';
+import { GraphQLDateTime } from 'graphql-scalars';
 
 import { AccountReferenceInput } from './AccountReferenceInput';
 
@@ -11,7 +10,7 @@ export const UpdateCreateInput = new GraphQLInputObjectType({
     title: { type: new GraphQLNonNull(GraphQLString) },
     isPrivate: { type: GraphQLBoolean },
     isChangelog: { type: GraphQLBoolean },
-    makePublicOn: { type: IsoDateString },
+    makePublicOn: { type: GraphQLDateTime },
     html: { type: new GraphQLNonNull(GraphQLString) },
     fromAccount: { type: AccountReferenceInput },
     account: { type: new GraphQLNonNull(AccountReferenceInput) },
